+++ .env.template
@@ -2,10 +2,11 @@ ROOT_DIR=
 MODEL_CACHE_DIR=
 
 ## Postgres
+POSTGRES_HOST=postgres
 POSTGRES_DB=opendatarepository
 POSTGRES_USER=opendatarepository
 POSTGRES_PASSWORD=opendatarepository
-POSTGRES_PORT=35432
+POSTGRES_PORT=5432
 TEST_POSTGRES_DB=test_opendatarepository
 
 # PGADMIN
