+++ docker-compose.yml
@@ -1,0 +1,51 @@
+name: omi-postgres
+services:
+  postgres:
+    image: pgvector/pgvector:pg16
+    environment:
+      POSTGRES_DB: ${POSTGRES_DB}
+      POSTGRES_USER: ${POSTGRES_USER}
+      POSTGRES_PASSWORD: ${POSTGRES_PASSWORD}
+    volumes:
+      - postgres_data:/var/lib/postgresql/data
+    ports:
+      - "${POSTGRES_PORT:-35432}:5432"
+    networks:
+      - omi-network
+
+  pgadmin:
+    image: dpage/pgadmin4
+    environment:
+      PGADMIN_DEFAULT_EMAIL: ${PGADMIN_DEFAULT_EMAIL:-admin@example.com}
+      PGADMIN_DEFAULT_PASSWORD: ${PGADMIN_DEFAULT_PASSWORD:-admin}
+      PGADMIN_LISTEN_ADDRESS: 0.0.0.0
+    volumes:
+      - pgadmin_data:/var/lib/pgadmin
+    ports:
+      - "${PGADMIN_PORT:-35050}:80"
+    networks:
+      - omi-network
+    depends_on:
+      - postgres
+
+  odr-api:
+    build:
+      context: .
+      dockerfile: modules/odr_api/docker/Dockerfile.api
+    env_file:
+      - .env
+    ports:
+      - "31100:31100"
+    networks:
+      - omi-network
+    depends_on:
+      - postgres
+
+volumes:
+  postgres_data:
+  pgadmin_data:
+
+networks:
+  omi-network:
+    external: true
+    driver: bridge
