+++ modules/odr_api/Taskfile.api.yml
@@ -1,0 +1,55 @@
+version: 3
+
+tasks:
+  start:
+    desc: Start the api server
+    cmds:
+      - cmd: powershell.exe -ExecutionPolicy ByPass ./modules/odr_api/scripts/run_server.ps1 -debug
+        platforms: [windows]
+      - cmd: ./modules/odr_api/scripts/start_server.sh --debug
+        platforms: [linux, darwin]
+      - echo 'Start server stub'
+
+  debug:
+    desc: Start the api server in debug mode
+    cmds:
+      - cmd: python -m debugpy --listen 0.0.0.0:5678 --wait-for-client ./modules/odr_api/odr_api/main.py
+        platforms: [windows, linux, darwin]
+  build:
+    dir: '{{.ROOT_DIR}}'
+    cmds:
+      - docker build -t odr-api -f modules/odr_api/docker/Dockerfile.api .
+
+  run:
+    desc: Run the odr-api Docker container
+    dir: '{{.ROOT_DIR}}'
+    cmds:
+      - >
+        docker run -d
+        --name odr-api
+        -p 31100:31100
+        --env-file .env
+        -e POSTGRES_URL={{.POSTGRES_URL | default "postgres"}}
+        odr-api
+    vars:
+      POSTGRES_URL:
+        sh: echo $POSTGRES_URL
+
+  stop:
+    cmds:
+      - docker stop odr-api
+      - docker rm odr-api
+
+  logs:
+    cmds:
+      - docker logs odr-api -f
+
+  shell:
+    cmds:
+      - docker exec -it odr-api /bin/sh
+
+  redeploy:
+    cmds:
+      - task: stop
+      - task: build
+      - task: run
